// libs/io-react-native-wallet/src/credential/presentation/api/types.ts
export type RequestObject = {
  client_id: string;
  client_metadata?: ClientMetadata;
  dcql_query: Record<string, unknown>;
  iss: string;
  nonce: string;
  response_mode: "direct_post.jwt";
  response_type: "vp_token";
  response_uri: string;
  state?: string;
  trust_chain?: string[];
  x5c?: string[];
};

type ClientMetadata = {
  client_name?: string;
  encrypted_response_enc_values_supported: string[];
  jwks: jsonWebKeySet;
  logo_uri?: string;
  vp_formats_supported: Record<string, { "sd-jwt_alg_values"?: string[] }>;
};